+++ lib/language_pack.rb
@@ -27,7 +27,10 @@ module LanguagePack
     warn_io = LanguagePack::ShellHelpers::WarnIO.new
     user_env_hash = LanguagePack::ShellHelpers.user_env_hash
     bundler_cache = LanguagePack::BundlerCache.new(cache, stack)
-    bundler_version = LanguagePack::Helpers::BundlerWrapper.detect_bundler_version(contents: gemfile_lock.contents)
+    bundler_version = LanguagePack::Helpers::BundlerWrapper.resolve_bundler_version(
+      warn_io: warn_io,
+      gemfile_lock: gemfile_lock,
+    )
 
     metadata = LanguagePack::Metadata.new(cache_path: cache_path)
     new_app = metadata.empty?
@@ -50,7 +53,7 @@ module LanguagePack
       io: warn_io
     )
 
-    bundler = Helpers::BundlerWrapper.new(bundler_path: ruby_version.bundler_directory).install
+    bundler = Helpers::BundlerWrapper.new(bundler_path: ruby_version.bundler_directory, bundler_version: bundler_version).install
     default_config_vars = Ruby.default_config_vars(metadata: metadata, ruby_version: ruby_version, bundler: bundler, environment_name: environment_name)
     Ruby.setup_language_pack_environment(
       app_path: app_path.expand_path,