+++ lib/language_pack/helpers/bundler_wrapper.rb
@@ -9,10 +9,10 @@ require "json"
 #
 # Example:
 #
-#   bundler = LanguagePack::Helpers::BundlerWrapper.new(bundler_path: "vendor/bundle/ruby/3.2.0")
+#   bundler = LanguagePack::Helpers::BundlerWrapper.new(bundler_path: "vendor/bundle/ruby/3.2.0", bundler_version: "2.5.7")
 #   bundler.install
-#   bundler.version                 => "1.15.2"
-#   bundler.dir_name                => "bundler-1.15.2"
+#   bundler.version                 => "2.5.23"
+#   bundler.dir_name                => "bundler-2.5.23"
 #   bundler.has_gem?("railties")    => true
 #   bundler.gem_version("railties") => "5.2.2"
 #   bundler.clean
@@ -22,94 +22,21 @@ require "json"
 # of an isolated dyno, you must call `BundlerWrapper#clean`. To reset the environment
 # variable:
 #
-#   bundler = LanguagePack::Helpers::BundlerWrapper.new(bundler_path: "vendor/bundle/ruby/3.2.0")
+#   bundler = LanguagePack::Helpers::BundlerWrapper.new(bundler_path: "vendor/bundle/ruby/3.2.0", bundler_version: "2.5.7")
 #   bundler.install
 #   bundler.clean # <========== IMPORTANT =============
 #
 class LanguagePack::Helpers::BundlerWrapper
   include LanguagePack::ShellHelpers
 
-  BLESSED_BUNDLER_VERSIONS = {}
   # Heroku-22's oldest Ruby version is 3.1
-  BLESSED_BUNDLER_VERSIONS["2.3"] = "2.3.25"
-  BLESSED_BUNDLER_VERSIONS["2.4"] = "2.4.22"
-  BLESSED_BUNDLER_VERSIONS["2.5"] = "2.5.23"
-  BLESSED_BUNDLER_VERSIONS["2.6"] = "2.6.9"
-  BLESSED_BUNDLER_VERSIONS["2.7"] = "2.7.2"
-  BLESSED_BUNDLER_VERSIONS["4.0"] = "4.0.0"
-
-  SORTED_KEYS = BLESSED_BUNDLER_VERSIONS.keys.map { |k| Gem::Version.new(k) }.sort
-  BUNDLER_2_SORTED_KEYS = SORTED_KEYS.select { |k| k.segments.first == 2 }
-  BUNDLER_2_SMALLEST = BUNDLER_2_SORTED_KEYS.first.to_s
-  BUNDLER_2_LARGEST = BUNDLER_2_SORTED_KEYS.last.to_s
-
-  BUNDLER_4_SORTED_KEYS = SORTED_KEYS.select { |k| k.segments.first == 4 }
-  BUNDLER_4_SMALLEST = BUNDLER_4_SORTED_KEYS.first.to_s
-  BUNDLER_4_LARGEST = BUNDLER_4_SORTED_KEYS.last.to_s
-
-  DEFAULT_VERSION = BLESSED_BUNDLER_VERSIONS["2.3"]
-
-  # Convert arbitrary `<Major>.<Minor>.x` versions
-  BLESSED_BUNDLER_VERSIONS.default_proc = Proc.new do |hash, key|
-    case Gem::Version.new(key).segments.first
-    when 4
-      hash[BUNDLER_4_LARGEST]
-    when 2
-      if Gem::Version.new(key) > Gem::Version.new(BUNDLER_2_LARGEST)
-        hash[BUNDLER_2_LARGEST]
-      elsif Gem::Version.new(key) < Gem::Version.new(BUNDLER_2_SMALLEST)
-        hash[BUNDLER_2_SMALLEST]
-      else
-        raise UnsupportedBundlerVersion.new(hash, key)
-      end
-    else
-      raise UnsupportedBundlerVersion.new(hash, key)
-    end
-  end
-
-  def self.detect_bundler_version(contents: , bundled_with: contents.match(BUNDLED_WITH_REGEX))
-    if bundled_with
-      major = bundled_with[:major]
-      minor = bundled_with[:minor]
-      version = BLESSED_BUNDLER_VERSIONS["#{major}.#{minor}"]
-      version
-    else
-      DEFAULT_VERSION
-    end
-  end
-
-  BUNDLED_WITH_REGEX = /^BUNDLED WITH$(\r?\n) {2,3}(?<version>(?<major>\d+)\.(?<minor>\d+)\.\d+)/m
-
-  class GemfileParseError < BuildpackError
-    def initialize(error)
-      msg = String.new("There was an error parsing your Gemfile, we cannot continue\n")
-      msg << error
-      super msg
-    end
-  end
-
-  class UnsupportedBundlerVersion < BuildpackError
-    def initialize(version_hash, major_minor)
-      msg = String.new("Your Gemfile.lock indicates you need bundler `#{major_minor}.x`\n")
-      msg << "which is not currently supported. You can deploy with bundler version:\n"
-      version_hash.keys.each do |v|
-        msg << "  - `#{v}.x`\n"
-      end
-      msg << "\nTo use another version of bundler, update your `Gemfile.lock` to point\n"
-      msg << "to a supported version. For example:\n"
-      msg << "\n"
-      msg << "```\n"
-      msg << "BUNDLED WITH\n"
-      msg << "   #{DEFAULT_VERSION}\n"
-      msg << "```\n"
-      super msg
-    end
-  end
+  DEFAULT_VERSION = "2.3.25"
 
   attr_reader :bundler_path
 
   def initialize(
       bundler_path:,
+      bundler_version:,
       gemfile_path: Pathname.new("./Gemfile"),
       report: HerokuBuildReport::GLOBAL
     )
@@ -117,21 +44,11 @@ class LanguagePack::Helpers::BundlerWrapper
     @gemfile_path         = gemfile_path
     @gemfile_lock_path    = Pathname.new("#{@gemfile_path}.lock")
 
-    contents = @gemfile_lock_path.read(mode: "rt")
-    bundled_with = contents.match(BUNDLED_WITH_REGEX)
     dot_ruby_version_file = @gemfile_lock_path.join("..").join(".ruby-version")
     @report.capture(
-      "bundler.bundled_with" => bundled_with&.[]("version") || "empty",
-      # We use this bundler class to detect the Requested ruby version from the Gemfile.lock
-      # Rails 8 stopped generating `RUBY VERSION` in the Gemfile.lock and started generating
-      # a `.ruby-version` file. This will observe the formats to help guide implementation
-      # decisions
       "ruby.dot_ruby_version" => dot_ruby_version_file.exist? ? dot_ruby_version_file.read&.strip : nil
     )
-    @version = self.class.detect_bundler_version(
-      contents: contents,
-      bundled_with: bundled_with
-    )
+    @version = bundler_version
     parts = @version.split(".")
     @report.capture(
       "bundler.version_installed" => @version,
@@ -175,28 +92,10 @@ class LanguagePack::Helpers::BundlerWrapper
     @dir_name
   end
 
-  def self.platform_to_version(bundle_platform_output)
-    if bundle_platform_output.match(/No ruby version specified/)
-      ""
-    else
-      bundle_platform_output.strip.sub('(', '').sub(')', '').sub(/(p-?\d+)/, ' \1').split.join('-')
-    end
-  end
-
-  def bundler_version_escape_valve!
-    topic("Removing BUNDLED WITH version in the Gemfile.lock")
-    contents = File.read(@gemfile_lock_path, mode: "rt")
-    File.open(@gemfile_lock_path, "w") do |f|
-      f.write contents.sub(/^BUNDLED WITH$(\r?\n) {2,3}(?<major>\d+)\.\d+\.\d+/m, '')
-    end
-  end
-
-  private
-  def fetch_bundler
+  private def fetch_bundler
     return true if Dir.exist?(bundler_path.join("gems", dir_name))
 
     topic("Installing bundler #{@version}")
-    bundler_version_escape_valve!
 
     # Install directory structure (as of Bundler 2.1.4):
     # - cache
@@ -211,7 +110,38 @@ class LanguagePack::Helpers::BundlerWrapper
   end
 
   # Runs a Ruby subprocess to parse the Gemfile.lock and return specs as a hash.
-  def specs_from_lockfile
+  private def specs_from_lockfile
     LanguagePack::Helpers::LockfileShellParser.call(lockfile_path: @gemfile_lock_path)
+  end
+
+  def self.resolve_bundler_version(gemfile_lock:, warn_io: )
+    version = gemfile_lock.bundler.version
+    if version
+      version
+    else
+      warn_io.warn(<<~WARNING)
+        Using default bundler version `#{DEFAULT_VERSION}`
+
+        The Ruby buildpack uses the `BUNDLED WITH` value in your `Gemfile.lock` to determine the version
+        of bundler to install. Your `Gemfile.lock` does not contain this section, so a default version
+        of bundler will be installed instead.
+
+        Heroku recommends that you have both a `RUBY VERSION` and `BUNDLED WITH` version listed in your `Gemfile.lock`.
+        You can add it to your project by running:
+
+        ```
+        $ bundle update --bundler
+        ```
+
+        Commit the results to git before redeploying:
+
+        ```
+        $ git add Gemfile.lock
+        $ git commit -m "Add BUNDLED WITH version"
+        ```
+      WARNING
+
+      DEFAULT_VERSION
+    end
   end
 end